Crate fedimint_gw_client

Modules

complete ๐

events

pay

Structs

GatewayClientContext

GatewayClientInit

GatewayClientModule

Client side Lightning module for the gateway.

Htlc

SwapParameters

Enums

GatewayClientStateMachines

GatewayExtPayStates

The high-level state of a reissue operation started with GatewayClientModule::gateway_pay_bolt11_invoice.

GatewayExtReceiveStates

The high-level state of an intercepted HTLC operation started with GatewayClientModule::gateway_handle_intercepted_htlc.

GatewayMeta

Traits

IGatewayClientV1

An interface between module implementation and the general Gateway